init
This is a slightly modified version of the test event JSON from the OMNIbus REST API docs.  I set First and Last Occurrence to zero in order to have the ObjectServer set to current time (note, this may only be the default when using the reference architecture triggers, which (in my not so humble opinion) everyone should use.  If you try this and end up with Jan 1st 1970 instead of the current date/time then you should change the zeros to the current epoch seconds.