added OpsTeam
Added OpsTeam, but I need to convert to the strings I use for the teams.  The schema has eight chars as the length of the OpsTeam field, and I am popping in 'Networking Team' in my testing, so it is truncated.  Have to look where else I use this field and see how it gets used and if I should be populating this directly, or if I should be populating a diff field.  There is a chance that I have a temporal trigger populating OpsTeam, in which case I should not be touching it directly and should be populating 'application' instead (I think this is the case, but need to look at it when I am fresh).